added unit tests and logging

+++ src/main/java/com/decode/decode/service/JournalEntryService.java
@@ -6,6 +6,8 @@ import com.decode.decode.entity.User;
 import com.decode.decode.repository.JournalEntryRepository;
 import lombok.extern.slf4j.Slf4j;
 import org.bson.types.ObjectId;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -24,10 +26,7 @@ public class JournalEntryService {
     @Autowired
     private UserService userService;
 
-    public void saveEntry(JournalEntry journalEntry) {
-            journalEntryRepository.save(journalEntry);
-    }
-
+    private static final Logger logger= LoggerFactory.getLogger(JournalEntryService.class);
 
     public void saveEntry(JournalEntry journalEntry,String userName) {
         try{
@@ -35,7 +34,7 @@ public class JournalEntryService {
             journalEntry.setDate(LocalDateTime.now());
             JournalEntry savedJournalEntry = journalEntryRepository.save(journalEntry);
             user.getJournalEntryList().add(savedJournalEntry);
-            userService.add(user);
+            userService.saveNewUser(user);
         }
         catch (Exception e){
             log.error("Exception"+ e);
@@ -52,10 +51,21 @@ public class JournalEntryService {
     }
 
     public void deleteById(ObjectId myid, String userName) {
-        User user = userService.findByUserName(userName);
-        user.getJournalEntryList().removeIf(x->x.getId().equals(myid));
-        userService.add(user);
-        journalEntryRepository.deleteById(myid);
+        try{
+            User user = userService.findByUserName(userName);
+            boolean removed = user.getJournalEntryList().removeIf(x -> x.getId().equals(myid));
+            if(removed) {
+                userService.saveUser(user);
+                journalEntryRepository.deleteById(myid);
+            }
+        }
+        catch(Exception e){
+            System.out.println(e);
+            throw new RuntimeException("An error occured while deleting an entry: " + e);
+        }
+    }
+    public List<JournalEntry> findUserName(String userName){
+        return null;
     }
 
 }
